+++ spec/components/emergency_banner_spec.rb
@@ -87,4 +87,13 @@ describe "Emergency Banner", type: :view do
     assert_select ".gem-c-emergency-banner[data-ga4-set-indexes]"
     assert_select ".gem-c-emergency-banner[data-ga4-link='{\"event_name\":\"navigation\",\"type\":\"emergency banner\",\"section\":\"His Royal Highness Henry VIII\"}']"
   end
+
+  it "renders banner without ga4 attributes" do
+    render_component(emergency_banner_attributes({ campaign_class: "notable-death", ga4_tracking: false }))
+    assert_no_selector ".gem-c-emergency-banner[data-ga4-emergency-banner]"
+    assert_no_selector ".gem-c-emergency-banner[data-module=ga4-link-tracker]"
+    assert_no_selector ".gem-c-emergency-banner[data-ga4-track-links-only]"
+    assert_no_selector ".gem-c-emergency-banner[data-ga4-set-indexes]"
+    assert_no_selector ".gem-c-emergency-banner[data-ga4-link]"
+  end
 end