+++ spec/javascripts/govuk_publishing_components/analytics-ga4/ga4-page-views.spec.js
@@ -52,7 +52,8 @@ describe('Google Tag Manager page view tracking', function () {
 
         dynamic: 'false',
         emergency_banner: undefined,
-        phase_banner: undefined
+        phase_banner: undefined,
+        devolved_nations_banner: undefined
       }
     }
     window.dataLayer = []
@@ -430,6 +431,16 @@ describe('Google Tag Manager page view tracking', function () {
     div.setAttribute('data-ga4-phase-banner', 'beta')
     document.body.appendChild(div)
     expected.page_view.phase_banner = 'beta'
+    GOVUK.analyticsGa4.analyticsModules.PageViewTracker.init()
+    expect(window.dataLayer[0]).toEqual(expected)
+    document.body.removeChild(div)
+  })
+
+  it('correctly sets the devolved_nations_banner parameter', function () {
+    var div = document.createElement('div')
+    div.setAttribute('data-ga4-devolved-nations-banner', 'England, Scotland, Wales')
+    document.body.appendChild(div)
+    expected.page_view.devolved_nations_banner = 'England, Scotland, Wales'
     GOVUK.analyticsGa4.analyticsModules.PageViewTracker.init()
     expect(window.dataLayer[0]).toEqual(expected)
     document.body.removeChild(div)